python/requirements.txt
Django==1.11.6
django-filter==1.1.0
djangorestframework==3.7.3
Markdown==2.6.9
python-decouple==3.1
pytz==2017.3
